Refactor and inprove state hoisting in BackoupSourceConfiguration #16
